Set requires_encoding in Custom parameter class to False
This fixes a bug in the contains_rdkit and contains_mordred properties of the searchspace, which would break otherwise. The new logic is that requires_encoding should indicate whether a parameter needs "further" encoding by BayBE, instead of indicating if the parameter already "is" naturally encoded (as is the case for Custom parameters).