Implement temporary workaround to get tests passing
To test that the code still runs after all the previous changes, a temporary workaround is implemented that allows to run the current tests without having to replace their config dicts. For this purpose, the `from_dict` method was temporarily added as a compatibility layer to use the new code structure in combination with the old dict format.