Define our own converter for serialization
The approach implemented in this commit still relies on a single monolithic converter for all default serialization tasks but no longer uses the cattrs.global_converter object. This is a step in the right direction as it avoids potential interference with other cattrs-dependent code outside the BayBE project scope.

For the optional remaining TODO see:
https://dev.azure.com/Uptimize/Bayesian%20Back%20End/_workitems/edit/19106