Drop "batches" mode from TwoPhaseStrategy
The reason is that it requires changing the signature of Strategy.recommend. While simply adding the batch count would suffice to implement the "batches" mode, this would mess up the interface since information about the training data would enter via three different arguments. A refactoring should be done as soon as a breaking change like this happens.