Fix bug in GP posterior computation
The `GaussianProcessSurrogate._posterior` method implements the
posterior computation where the input is already scale-transforemd.
This should **not** be passed to botorch, because the transition
point to botorch is on a layer where the concept of scaling
is irrelevant / does not exist (scaling is a purely surrogate-
internal process). Instead, we go via the adapter model, which wraps
the appropriate method.