Change plan and expose internal GP, but with disabled base class scaling
The problem is that qNIPV requires a model that can `fantasize`, which
our `AdapterModel` is not capable of. We thus expose the actual botorch
GP model instead. But for this to work correctly, we still need to
ensure that scaling happens **after** the transition point to
`optimize_acqf_*`:

`optimize_acqf_*` is called with **unscaled** parameters in **computational
representation**, so it must be provided a surrogate that operates on
the same domain. In regular operation, the `AdapterModel` takes care
of this by exposing the method `posterior_comp_rep` as the correct entry
point. When we now expose the internal botorch GP model, the exposed
posterior method was calibrated on the already scaled training data,
resulting in a mismatch. The workaround is to deactive the base class
scaling and instead let the GP internally do the scaling work.