Sort all discrete parameters
If unsorted, this can cause problems with reproducibility in the sense
that the same parameter content provided in a different order can lead
to different optimization results. For instance, the RandomRecommender
randomly selects rows from the comp_rep dataframe for the discrete
subspace, which will be ordered differently if the parameter values come
in a different order. This can lead to rather surprising behavior, like
in situations when the parameter values are given as the output of
Python's set function, whose order depends on PYTHONHASHSEED.