Fix installation if scm is not available.
In the specific case, baybe was installed format
the git repository inside a docker container that
did not provide the ` git` command, leading
setuptools-scm to fail. Same happens if for
any reason the .git directory is missing (e.g.
when downloading a tarbal from github).

The fallback version is set to `0.0+unknown` to
ensure we adhere to PyPa (https://packaging.python.org/en/latest/specifications/version-specifiers/#examples-of-compliant-version-schemes
) while still ensuring any user will be able to
recognize this as a non-standard version
identifier.