Fix guard clause
The wrong type check resulted in `PosteriorStandardDeviation` always being minimized (regardless of its own `maximize` flag) when used in combination with a single `MIN` target, because the flag was overwritten in the code lines below. The other two active learning acquisition functions were not affected since the code execution does not reach the corresponding line in these cases.