Drop unnecessary xfail
The xfail was introduced because qNIPV does not support (BoTorch)
objectives, but the new target transformation framework now has separate
exection paths for affine and non-affine transformations, making the
xfail obsolete. In the former case, which applies for the given test
case, transformations are handled via BoTorch's posterior transformation
mechanism. Hence, we do not run into any limitations of the qNIPV
acquisition function.